Renamed Test to Utils. Moved away some of the functionality to ETS
