+++ lib/utils.ex
@@ -1,30 +1,9 @@
-defmodule MerklePatriciaTree.Test do
+defmodule MerklePatriciaTree.Utils do
   @moduledoc """
-  Helper functions related to creating a MerklePatriciaTree
-  database in :ets or :leveldb. This is to be used when
-  you need to have a persisted table for a test case, which
-  is basically any test that involves these tables. The tables
-  generated by this module are supposed to be exceedingly temporary.
+  Helper functions related to common operations
   """
 
   @hash_bytes 32
-
-  @doc """
-  Returns a random :ets database suitable for testing
-
-  ## Examples
-
-      iex> {MerklePatriciaTree.DB.ETS, db_ref} = MerklePatriciaTree.Test.random_ets_db()
-      iex> :ets.info(db_ref)[:type]
-      :set
-
-      iex> {MerklePatriciaTree.DB.ETS, db_ref} = MerklePatriciaTree.Test.random_ets_db(:test1)
-      iex> :ets.info(db_ref)[:name]
-      :test1
-  """
-  def random_ets_db(name \\ nil) do
-    MerklePatriciaTree.DB.ETS.init(name || MerklePatriciaTree.Test.random_atom(20))
-  end
 
   @doc """
   Returns a semi-random string of length `length` that
